Move the fee to separate file
